Refactor: extracted some methods #5
